requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
flask-mongoengine-3==1.0.9
Flask-WTF==1.2.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.42
langchain-openai==0.1.3
langsmith==0.1.46
MarkupSafe==2.1.5
mongoengine==0.28.2
openai==1.17.1
orjson==3.10.0
packaging==23.2
pydantic==2.7.0
pydantic_core==2.18.1
pymongo==4.6.3
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
WTForms==3.1.2
